chore(update for ivy): make example work with ivy (closes #2)
Update dependencies, config, minor code changes to work with Ivy.
