+++ .vscode/launch.json
@@ -3,11 +3,14 @@
   // Hover to view descriptions of existing attributes.
   // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
   "version": "0.2.0",
-  "configurations": [{
-    "type": "chrome",
-    "request": "launch",
-    "name": "Launch Chrome against localhost",
-    "url": "http://localhost:4200",
-    "webRoot": "${workspaceFolder}"
-  }]
+  "configurations": [
+    {
+      "name": "Launch Edge against localhost (ng)",
+      "type": "pwa-msedge",
+      "request": "launch",
+      "url": "http://localhost:4200",
+      "webRoot": "${workspaceFolder}",
+      "showAsyncStacks": true
+    }
+  ]
 }
