+++ package.json
@@ -9,29 +9,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.1.0",
-    "@angular/cdk": "~7.1.1",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/material": "^7.1.1",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
-    "core-js": "^2.5.4",
-    "rxjs": "~6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "@angular/animations": "~12.2.1",
+    "@angular/cdk": "~12.2.1",
+    "@angular/common": "~12.2.1",
+    "@angular/compiler": "~12.2.1",
+    "@angular/core": "~12.2.1",
+    "@angular/forms": "~12.2.1",
+    "@angular/material": "^12.2.1",
+    "@angular/platform-browser": "~12.2.1",
+    "@angular/platform-browser-dynamic": "~12.2.1",
+    "@angular/router": "~12.2.1",
+    "core-js": "^3.16.1",
+    "rxjs": "~7.3.0",
+    "tslib": "^2.3.1",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular/cli": "~7.1.2",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
-    "@types/node": "~8.9.4",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "@angular-devkit/build-angular": "~12.2.1",
+    "@angular/cli": "~12.2.1",
+    "@angular/compiler-cli": "~12.2.1",
+    "@angular/language-service": "~12.2.1",
+    "@types/node": "~16.6.1",
+    "ts-node": "~10.2.0",
+    "tslint": "~6.1.3",
+    "typescript": "~4.3.5"
   }
 }
