+++ src/app/app.component.html
@@ -2,7 +2,6 @@
   <h2>Template Examples</h2>
 </mat-toolbar>
 
-
 <section *ngFor="let example of examples?.values()">
   <h5>{{example.id}}</h5>
   <div class="example">
